For #4261: Move AutoFitTextureView to feature-qr module
