+++ components/feature/qr/src/test/java/mozilla/components/feature/qr/views/AutoFitTextureViewTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.support.base.android.view
+package mozilla.components.feature.qr.views
 
 import android.view.View
 import androidx.test.core.app.ApplicationProvider
